split schema from server part

+++ bash/clean.bash
@@ -1,3 +1,5 @@
-rm -rf ./node_modules;
-rm -rf ./.gitignore;
-rm -rf ./.npmignore;
+if [ -d "./node_modules" ]; then
+  rm -rf ./node_modules;
+  rm -rf ./.gitignore;
+  rm -rf ./.npmignore;
+fi
